requirements.txt
beartype==0.19.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coloredlogs==15.0.1
cryptography==44.0.0
filelock==3.17.0
flatbuffers==25.1.21
fsspec==2024.12.0
huggingface-hub==0.25.0
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.2
onnxruntime==1.19.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
protobuf==5.29.3
pyclipper==1.3.0.post5
pycparser==2.22
pypdfium2==4.30.1
PyYAML==6.0.2
requests==2.32.3
shapely==2.0.6
six==1.17.0
sympy==1.13.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
